pyproject.toml
[project]
name = "ustudyzakaztzproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.23.0",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
]
